# Optimize Key and KeyPtr primitives