--- a/primitives/benches/bench.rs
+++ b/primitives/benches/bench.rs
@@ -12,19 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use criterion::{
+    black_box,
+    criterion_group,
+    criterion_main,
+    Criterion,
+};
 use ink_primitives::{
     Key,
     Key2,
     KeyPtr,
     KeyPtr2,
-};
-use criterion::{
-    black_box,
-    criterion_group,
-    criterion_main,
-    // BatchSize,
-    // BenchmarkId,
-    Criterion,
 };
 
 criterion_group!(
@@ -43,102 +41,116 @@ criterion_group!(
     bench_key_ptr_advance_by_repeat,
     bench_key_ptr2_advance_by_repeat,
 );
-criterion_main!(
-    bench_key,
-    bench_key_ptr,
-);
+criterion_main!(bench_key, bench_key_ptr,);
 
 fn bench_key_add_assign_u64(c: &mut Criterion) {
     let key = Key([0x00; 32]);
-    c.bench_function("Key::add_assign(u64)", |b| b.iter(|| {
-        let mut copy = black_box(key);
-        let _ = black_box(copy += 1u64);
-    }));
+    c.bench_function("Key::add_assign(u64)", |b| {
+        b.iter(|| {
+            let mut copy = black_box(key);
+            let _ = black_box(copy += 1u64);
+        })
+    });
 }
 
 fn bench_key2_add_assign_u64(c: &mut Criterion) {
     let key = Key2::from([0x00; 32]);
-    c.bench_function("Key2::add_assign(u64)", |b| b.iter(|| {
-        let mut copy = black_box(key);
-        let _ = black_box(copy += 1u64);
-    }));
+    c.bench_function("Key2::add_assign(u64)", |b| {
+        b.iter(|| {
+            let mut copy = black_box(key);
+            let _ = black_box(copy += 1u64);
+        })
+    });
 }
 
 fn bench_key2_add_assign_u64_one_ofvl(c: &mut Criterion) {
     let key = Key2::from([
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
     ]);
-    c.bench_function("Key2::add_assign(u64) - 1 ofvl", |b| b.iter(|| {
-        let mut copy = black_box(key);
-        let _ = black_box(copy += 1u64);
-    }));
+    c.bench_function("Key2::add_assign(u64) - 1 ofvl", |b| {
+        b.iter(|| {
+            let mut copy = black_box(key);
+            let _ = black_box(copy += 1u64);
+        })
+    });
 }
 
 fn bench_key2_add_assign_u64_two_ofvls(c: &mut Criterion) {
     let key = Key2::from([
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
+        0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
     ]);
-    c.bench_function("Key2::add_assign(u64) - 2 ofvls", |b| b.iter(|| {
-        let mut copy = black_box(key);
-        let _ = black_box(copy += 1u64);
-    }));
+    c.bench_function("Key2::add_assign(u64) - 2 ofvls", |b| {
+        b.iter(|| {
+            let mut copy = black_box(key);
+            let _ = black_box(copy += 1u64);
+        })
+    });
 }
 
 fn bench_key2_add_assign_u64_three_ofvls(c: &mut Criterion) {
     let key = Key2::from([
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
-        0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF,
+        0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00,
+        0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
     ]);
-    c.bench_function("Key2::add_assign(u64) - 3 ofvls", |b| b.iter(|| {
-        let mut copy = black_box(key);
-        let _ = black_box(copy += 1u64);
-    }));
+    c.bench_function("Key2::add_assign(u64) - 3 ofvls", |b| {
+        b.iter(|| {
+            let mut copy = black_box(key);
+            let _ = black_box(copy += 1u64);
+        })
+    });
 }
 
 fn bench_key2_add_assign_u64_wrap(c: &mut Criterion) {
     let key = Key2::from([0xFF; 32]);
-    c.bench_function("Key2::add_assign(u64) - wrap", |b| b.iter(|| {
-        let mut copy = black_box(key);
-        let _ = black_box(copy += 1u64);
-    }));
+    c.bench_function("Key2::add_assign(u64) - wrap", |b| {
+        b.iter(|| {
+            let mut copy = black_box(key);
+            let _ = black_box(copy += 1u64);
+        })
+    });
 }
 
 fn bench_key_ptr_advance_by(c: &mut Criterion) {
     let key = Key([0x00; 32]);
-    c.bench_function("KeyPtr::advance_by copy", |b| b.iter(|| {
-        let mut key_ptr = KeyPtr::from(key.clone());
-        let _ = black_box(key_ptr.advance_by(1));
-    }));
+    c.bench_function("KeyPtr::advance_by copy", |b| {
+        b.iter(|| {
+            let mut key_ptr = KeyPtr::from(key.clone());
+            let _ = black_box(key_ptr.advance_by(1));
+        })
+    });
 }
 
 fn bench_key_ptr2_advance_by(c: &mut Criterion) {
     let key = Key2::from([0x00; 32]);
-    c.bench_function("KeyPtr2::advance_by copy", |b| b.iter(|| {
-        let mut key_ptr = KeyPtr2::from(key.clone());
-        let _ = black_box(key_ptr.advance_by(1));
-    }));
+    c.bench_function("KeyPtr2::advance_by copy", |b| {
+        b.iter(|| {
+            let mut key_ptr = KeyPtr2::from(key.clone());
+            let _ = black_box(key_ptr.advance_by(1));
+        })
+    });
 }
 
 fn bench_key_ptr_advance_by_repeat(c: &mut Criterion) {
     let key = Key([0x00; 32]);
     let mut key_ptr = KeyPtr::from(key.clone());
-    c.bench_function("KeyPtr::advance_by reuse", |b| b.iter(|| {
-        let _ = black_box(key_ptr.advance_by(1));
-    }));
+    c.bench_function("KeyPtr::advance_by reuse", |b| {
+        b.iter(|| {
+            let _ = black_box(key_ptr.advance_by(1));
+        })
+    });
 }
 
 fn bench_key_ptr2_advance_by_repeat(c: &mut Criterion) {
     let key = Key2::from([0x00; 32]);
     let mut key_ptr = KeyPtr2::from(key.clone());
-    c.bench_function("KeyPtr2::advance_by reuse", |b| b.iter(|| {
-        let _ = black_box(key_ptr.advance_by(1));
-    }));
+    c.bench_function("KeyPtr2::advance_by reuse", |b| {
+        b.iter(|| {
+            let _ = black_box(key_ptr.advance_by(1));
+        })
+    });
 }