--- a/primitives/src/lib.rs
+++ b/primitives/src/lib.rs
@@ -26,8 +26,16 @@
 mod byte_utils;
 pub mod hash;
 mod key;
+mod key2;
+mod key_ptr;
 
-pub use self::key::{
-    Key,
-    KeyDiff,
+pub use self::{
+    key::{
+        Key,
+        KeyDiff,
+    },
+    key2::{
+        Key as Key2,
+    },
+    key_ptr::KeyPtr,
 };