--- a/primitives/src/lib.rs
+++ b/primitives/src/lib.rs
@@ -28,6 +28,7 @@ pub mod hash;
 mod key;
 mod key2;
 mod key_ptr;
+mod key_ptr2;
 
 pub use self::{
     key::{
@@ -38,4 +39,5 @@ pub use self::{
         Key as Key2,
     },
     key_ptr::KeyPtr,
+    key_ptr2::KeyPtr as KeyPtr2,
 };