Benchmarks between the old and new `Key` and `KeyPtr` primitives below:

tl;dr:

- This PR does _NOT_ integrate the new `Key` and `KeyPtr` abstractions into the rest of ink!. A follow-up PR is going to do just that after #422 is merged.
- `Key2::add_assign` is approx. 1600% more efficient:
    - Drops from 24ns to 1.5ns.
- `KeyPtr2::advance_by` is approx. 1900% more efficient.
    - Drops from 12ns to 640ps.
- Tested on local machine, not via Wasm executor. However, we estimate a strong correlation between native and Wasm execution performance.

```
     Running /home/robrob/projects/ink/target/release/deps/bench-e4765ced662280c7
Key::add_assign(u64)    time:   [23.987 ns 24.077 ns 24.181 ns]                                  
                        change: [-0.8701% -0.0716% +0.8534%] (p = 0.87 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

Key2::add_assign(u64)   time:   [1.4984 ns 1.5050 ns 1.5122 ns]                                   
                        change: [-0.6685% -0.0464% +0.5931%] (p = 0.90 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe

Key2::add_assign(u64) - 1 ofvl                                                                             
                        time:   [1.5038 ns 1.5086 ns 1.5136 ns]
                        change: [-1.1945% -0.5885% +0.0526%] (p = 0.07 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe

Key2::add_assign(u64) - 2 ofvls                                                                             
                        time:   [1.4900 ns 1.4953 ns 1.5011 ns]
                        change: [-1.2463% -0.6133% +0.0815%] (p = 0.07 > 0.05)
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

Key2::add_assign(u64) - 3 ofvls                                                                             
                        time:   [1.4966 ns 1.5025 ns 1.5087 ns]
                        change: [-1.2574% -0.5210% +0.2675%] (p = 0.20 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) high mild
  4 (4.00%) high severe

Key2::add_assign(u64) - wrap                                                                             
                        time:   [1.5052 ns 1.5114 ns 1.5183 ns]
                        change: [+0.1999% +1.1495% +2.3522%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 11 outliers among 100 measurements (11.00%)
  7 (7.00%) high mild
  4 (4.00%) high severe

KeyPtr::advance_by      time:   [12.681 ns 12.736 ns 12.794 ns]                                
                        change: [-1.7342% -0.8787% -0.1030%] (p = 0.04 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

KeyPtr2::advance_by     time:   [315.05 ps 316.28 ps 317.69 ps]                                
                        change: [-1.5724% -0.7687% +0.0034%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) low severe
  1 (1.00%) low mild
  6 (6.00%) high mild
  2 (2.00%) high severe

KeyPtr::advance_by #2   time:   [12.056 ns 12.101 ns 12.150 ns]                                   
                        change: [-0.8768% -0.3147% +0.2457%] (p = 0.27 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild

KeyPtr2::advance_by #2  time:   [627.80 ps 630.46 ps 633.24 ps]                                    
                        change: [-0.3705% +0.3682% +1.0752%] (p = 0.35 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
```